// src/api/sqlmodule/model/dbmsDictionaryResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export interface DbmsDictionaryResponse {
  id?: string;
  dbmsName?: string;
  dbmsSystemName?: string;
  dockerImage?: string;
  defaultPort?: number;
  envUserKey?: string;
  envPasswordKey?: string;
  envDatabaseKey?: string;
  /** @nullable */
  extraEnvConfig?: string | null;
  defaultDatabase?: string;
  defaultUsername?: string;
  createdById?: string;
  createdAt?: string;
  updatedById?: string;
  updatedAt?: string;
  isActive?: boolean;
  isAvailable?: boolean;
  /** @nullable */
  unavailableReason?: string | null;
  canManageCatalog?: boolean;
}