// src/api/sqlmodule/model/targetDbResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

/**
 * Данные целевой БД-песочницы.
 */
export interface TargetDbResponse {
  /** Уникальный идентификатор записи. */
  id?: string;
  /** Идентификатор СУБД из справочника. */
  dbmsId?: string;
  /** Имя БД-песочницы. */
  dbName?: string;
  /**
     * Описание назначения БД (может быть `null`).
     * @nullable
     */
  description?: string | null;
  /** Признак доступа только на чтение. */
  isReadOnly?: boolean;
  /** Идентификатор пользователя, создавшего запись. */
  createdById?: string;
  /** Дата и время создания записи (UTC). */
  createdAt?: string;
  /** Идентификатор пользователя, последним изменившего запись. */
  updatedById?: string;
  /** Дата и время последнего изменения записи (UTC). */
  updatedAt?: string;
  /**
     * Отображаемое имя автора на момент создания.
     * @nullable
     */
  createdByName?: string | null;
  /**
     * Отображаемое имя последнего редактора.
     * @nullable
     */
  updatedByName?: string | null;
}